Move Virtualization controllers and Json in their own packages
The Virtualization-related controllers and JSon classes are starting to
be rather numerous. This commit is gathering them in a sub-package of
com.susr.manager.webui.controller.
