+++ java/code/src/com/suse/manager/webui/controllers/virtualization/VirtualGuestsController.java
@@ -12,7 +12,7 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package com.suse.manager.webui.controllers;
+package com.suse.manager.webui.controllers.virtualization;
 
 import static com.suse.manager.webui.utils.SparkApplicationHelper.json;
 import static com.suse.manager.webui.utils.SparkApplicationHelper.withCsrfToken;
@@ -54,13 +54,15 @@ import com.suse.manager.reactor.utils.OptionalTypeAdapterFactory;
 import com.suse.manager.virtualization.DomainCapabilitiesJson;
 import com.suse.manager.virtualization.GuestDefinition;
 import com.suse.manager.virtualization.HostCapabilitiesJson;
+import com.suse.manager.webui.controllers.ECMAScriptDateAdapter;
+import com.suse.manager.webui.controllers.MinionController;
+import com.suse.manager.webui.controllers.virtualization.gson.VirtualGuestSetterActionJson;
+import com.suse.manager.webui.controllers.virtualization.gson.VirtualGuestsBaseActionJson;
+import com.suse.manager.webui.controllers.virtualization.gson.VirtualGuestsUpdateActionJson;
 import com.suse.manager.webui.errors.NotFoundException;
 import com.suse.manager.webui.services.iface.VirtManager;
 import com.suse.manager.webui.utils.MinionActionUtils;
 import com.suse.manager.webui.utils.WebSockifyTokenBuilder;
-import com.suse.manager.webui.utils.gson.VirtualGuestSetterActionJson;
-import com.suse.manager.webui.utils.gson.VirtualGuestsBaseActionJson;
-import com.suse.manager.webui.utils.gson.VirtualGuestsUpdateActionJson;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.http.HttpStatus;
