+++ java/code/src/com/suse/manager/webui/controllers/virtualization/gson/VirtualVolumeBaseActionJson.java
@@ -12,7 +12,9 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package com.suse.manager.webui.utils.gson;
+package com.suse.manager.webui.controllers.virtualization.gson;
+
+import com.suse.manager.webui.utils.gson.ScheduledRequestJson;
 
 import java.util.List;
 import java.util.Map;
