+++ java/code/src/com/suse/manager/webui/controllers/virtualization/test/VirtualGuestsControllerTest.java
@@ -13,9 +13,7 @@
  * in this software or its documentation.
  */
 
-package com.suse.manager.webui.controllers.test;
-
-import static junit.framework.Assert.assertEquals;
+package com.suse.manager.webui.controllers.virtualization.test;
 
 import com.redhat.rhn.common.db.datasource.DataResult;
 import com.redhat.rhn.domain.action.Action;
@@ -42,14 +40,21 @@ import com.google.gson.reflect.TypeToken;
 import com.suse.manager.reactor.messaging.test.SaltTestUtils;
 import com.suse.manager.virtualization.DomainCapabilitiesJson;
 import com.suse.manager.virtualization.GuestDefinition;
-import com.suse.manager.webui.controllers.VirtualGuestsController;
 import com.suse.manager.virtualization.test.TestVirtManager;
+import com.suse.manager.webui.controllers.test.BaseControllerTestCase;
+import com.suse.manager.webui.controllers.virtualization.VirtualGuestsController;
 import com.suse.manager.webui.services.iface.VirtManager;
 import com.suse.manager.webui.services.impl.SaltService;
 
 import org.jmock.Expectations;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 import spark.HaltException;
 
@@ -86,7 +91,7 @@ public class VirtualGuestsControllerTest extends BaseControllerTestCase {
             @Override
             public Optional<Map<String, JsonElement>> getCapabilities(String minionId) {
                 return SaltTestUtils.getSaltResponse(
-                        "/com/suse/manager/webui/controllers/test/virt.guest.allcaps.json", null,
+                        "/com/suse/manager/webui/controllers/virtualization/test/virt.guest.allcaps.json", null,
                         new TypeToken<Map<String, JsonElement>>() { });
             }
 
@@ -156,7 +161,6 @@ public class VirtualGuestsControllerTest extends BaseControllerTestCase {
      *
      * @throws Exception if anything unexpected happens during the test
      */
-    @SuppressWarnings("unchecked")
     public void testStateChangeAction() throws Exception {
         VirtualInstance guest = host.getGuests().iterator().next();
         Long sid = host.getId();
@@ -187,7 +191,6 @@ public class VirtualGuestsControllerTest extends BaseControllerTestCase {
      *
      * @throws Exception if anything unexpected happens during the test
      */
-    @SuppressWarnings("unchecked")
     public void testSetVcpuAction() throws Exception {
         VirtualInstance guest = host.getGuests().iterator().next();
         Long sid = host.getId();
@@ -241,7 +244,6 @@ public class VirtualGuestsControllerTest extends BaseControllerTestCase {
      *
      * @throws Exception if anything unexpected happens during the test
      */
-    @SuppressWarnings("unchecked")
     public void testSetMemMultiAction() throws Exception {
 
         VirtualInstance[] guests = host.getGuests().toArray(new VirtualInstance[host.getGuests().size()]);
@@ -287,7 +289,6 @@ public class VirtualGuestsControllerTest extends BaseControllerTestCase {
      *
      * @throws Exception if anything unexpected happens during the test
      */
-    @SuppressWarnings("unchecked")
     public void testGetGuest() throws Exception {
         String json = virtualGuestsController.getGuest(
                 getRequestWithCsrf("/manager/api/systems/details/virtualization/guests/:sid/guest/:uuid",
