+++ java/code/src/com/suse/manager/webui/controllers/virtualization/test/VirtualPoolsControllerTest.java
@@ -12,19 +12,19 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package com.suse.manager.webui.controllers.test;
+package com.suse.manager.webui.controllers.virtualization.test;
 
 import static org.hamcrest.collection.IsIterableContainingInAnyOrder.containsInAnyOrder;
 
 import com.redhat.rhn.common.db.datasource.DataResult;
 import com.redhat.rhn.domain.action.Action;
 import com.redhat.rhn.domain.action.ActionFactory;
+import com.redhat.rhn.domain.action.virtualization.BaseVirtualizationVolumeAction;
 import com.redhat.rhn.domain.action.virtualization.VirtualizationPoolDeleteAction;
 import com.redhat.rhn.domain.action.virtualization.VirtualizationPoolRefreshAction;
-import com.redhat.rhn.domain.server.MinionServer;
 import com.redhat.rhn.domain.action.virtualization.VirtualizationPoolStartAction;
 import com.redhat.rhn.domain.action.virtualization.VirtualizationPoolStopAction;
-import com.redhat.rhn.domain.action.virtualization.BaseVirtualizationVolumeAction;
+import com.redhat.rhn.domain.server.MinionServer;
 import com.redhat.rhn.domain.server.Server;
 import com.redhat.rhn.frontend.context.Context;
 import com.redhat.rhn.frontend.dto.ScheduledAction;
@@ -43,11 +43,12 @@ import com.google.gson.reflect.TypeToken;
 import com.suse.manager.reactor.messaging.test.SaltTestUtils;
 import com.suse.manager.virtualization.PoolCapabilitiesJson;
 import com.suse.manager.virtualization.PoolCapabilitiesJson.PoolType;
-import com.suse.manager.webui.controllers.VirtualPoolsController;
 import com.suse.manager.virtualization.test.TestVirtManager;
+import com.suse.manager.webui.controllers.test.BaseControllerTestCase;
+import com.suse.manager.webui.controllers.virtualization.VirtualPoolsController;
+import com.suse.manager.webui.controllers.virtualization.gson.VirtualStoragePoolInfoJson;
 import com.suse.manager.webui.services.iface.VirtManager;
 import com.suse.manager.webui.services.impl.SaltService;
-import com.suse.manager.webui.utils.gson.VirtualStoragePoolInfoJson;
 
 import org.hamcrest.collection.IsMapContaining;
 import org.jmock.Expectations;
@@ -90,7 +91,7 @@ public class VirtualPoolsControllerTest extends BaseControllerTestCase {
             @Override
             public Map<String, JsonObject> getPools(String minionId) {
                 return SaltTestUtils.getSaltResponse(
-                        "/com/suse/manager/webui/controllers/test/virt.pool.info.json",
+                        "/com/suse/manager/webui/controllers/virtualization/test/virt.pool.info.json",
                         null,
                         new TypeToken<Map<String, JsonObject>>() { }).get();
             }
@@ -98,7 +99,7 @@ public class VirtualPoolsControllerTest extends BaseControllerTestCase {
             @Override
             public Map<String, Map<String, JsonObject>> getVolumes(String minionId) {
                 return SaltTestUtils.getSaltResponse(
-                        "/com/suse/manager/webui/controllers/test/virt.volume.info.json",
+                        "/com/suse/manager/webui/controllers/virtualization/test/virt.volume.info.json",
                         null,
                         new TypeToken<Map<String, Map<String, JsonObject>>>() { }).get();
             }
@@ -106,7 +107,7 @@ public class VirtualPoolsControllerTest extends BaseControllerTestCase {
             @Override
             public Optional<PoolCapabilitiesJson> getPoolCapabilities(String minionId) {
                 return SaltTestUtils.getSaltResponse(
-                        "/com/suse/manager/webui/controllers/test/virt.pool.caps.json",
+                        "/com/suse/manager/webui/controllers/virtualization/test/virt.pool.caps.json",
                         null,
                         new TypeToken<PoolCapabilitiesJson>() { });
             }
