+++ java/code/src/com/suse/manager/webui/controllers/virtualization/test/virt.pool.caps.json
